# openapi.v1.yaml
openapi: 3.0.3
info:
  title: datafusion-server
  description:
    This is a datafusion-server based on the OpenAPI 3.0 specification.
  version: 1.0.0
...
paths:
...
  /session/{sessionId}:
    get:
      tags:
        - Session
      summary: Get a existing session information
      operationId: session
      parameters:
        - name: sessionId
          in: path
          description: Session identifier
          required: true
          schema:
            type: string
          example: 29e7b63a-96e8-450a-8c23-6e43e1ec8337
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: 29e7b63a-96e8-450a-8c23-6e43e1ec8337
                  created:
                    type: string
                    example: 2023-01-22T05:25:01.626Z
                    description: Created at (RFC3339, UTC, milli-sec)
                  ttl:
                    type: number
                    example: 3547521
                    description: Remain time to live (sec)
        '400':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
    delete:
      tags:
        - Session
      summary: Remove a existing session
      operationId: removeSession
      parameters:
        - name: sessionId
          in: path
          description: Session identifier
          required: true
          schema:
            type: string
          example: 29e7b63a-96e8-450a-8c23-6e43e1ec8337
      responses:
        '200':
          description: Successful operation
        '400':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
        '500':
          description: Can not remove a existing session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'
...
  /session/{sessionId}/processor:
    post:
      tags:
        - Processor
      summary: Executes built-in processors
      operationId: processor
      parameters:
        - name: sessionId
          in: path
          description: Session identifier
          required: true
          schema:
            type: string
          example: 29e7b63a-96e8-450a-8c23-6e43e1ec8337
        - name: Request Body
          in: body
          description: Target data source
          required: true
          schema:
            $ref: '#/components/schemas/processor'
      responses:
        '204':
          description: Successful operation (No Content)
        '404':
          description: Session or data source not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'
        '400':
          description: Can not executes one or more processors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400'
...
components:
  schemas:
...
    error400:
      type: object
      properties:
        code:
          type: integer
          format: int16
          description: HTTP response code
          example: 400
        error:
          type: string
          description: Error identifier
          example: 'unsupported_data_source_type'
        message:
          type: string
          description: Detailed error message
          example: 'Unsupported data source type: xml'
    error404:
      type: object
      properties:
        code:
          type: integer
          format: int16
          description: HTTP response code
          example: 404
        error:
          type: string
          description: Error identifier
          example: 'session_not_found'
        message:
          type: string
          description: Detailed error message
          example: 'Incorrect session identifier'
    error500:
      type: object
      properties:
        code:
          type: integer
          format: int16
          description: HTTP response code
          example: 500
        error:
          type: string
          description: Error identifier
          example: 'io_error'
        message:
          type: string
          description: Detailed error message
          example: 'Failed to create data source into record batches'
...
    processor:
      type: object
      properties:
        mergeProcessors:
          type: array
          items:
            type: object
            properties:
              direction:
                type: string
                enum:
                  - row
                  - column
                example: row
                description: Merge direction
              baseTable:
                type: string
                example: table1
                description: Merge all target tables into the base table
              targetTables:
                type: array
                items:
                  type: string
                  example: ['table2', 'table3']
                  description: Merge target tables (Only for row direction)
              targets:
                type: array
                items:
                  type: object
                  properties:
                    table:
                      type: string
                      example: 'table2'
                    baseKeys:
                      type: array
                      items:
                        type: string
                        example: ['col1', 'col2']
                    targetKeys:
                      type: array
                      items:
                        type: string
                        example: ['id', 'category']
                description: Merge target tables by keys (Only for column direction)
              options:
                type: object
                properties:
                  distinct:
                    type: boolean
                    example: false
                    description: Removes same values records (Only for row direction)
                  removeAfterMerged:
                    type: boolean
                    example: false
                    description: Removes target tables when merge process has done
...